Added configs for X11 and bash

+++ install.sh
@@ -2,23 +2,35 @@
 
 BASEDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-#vim
+read -p "This will overwrite any existing config, are you sure? " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	#vim
 
-ln -sfv ${BASEDIR}/vim/vimrc ~/.vimrc
+	ln -sfv ${BASEDIR}/vim/vimrc ~/.vimrc
 
-rm -rf ~/.vim
-git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+	rm -rf ~/.vim
+	git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 
-vim +PluginInstall
+	vim +PluginInstall
 
-#tmux
+	#tmux
 
-ln -sfv ${BASEDIR}/tmux/tmux.conf ~/.tmux.conf
+	ln -sfv ${BASEDIR}/tmux/tmux.conf ~/.tmux.conf
 
-#i3
+	#i3
 
-ln -sfv ${BASEDIR}/i3/config ~/.i3/config
-ln -sfv ${BASEDIR}/i3/i3blocks.conf ~/.i3/i3blocks.conf
+	ln -sfv ${BASEDIR}/i3/config ~/.config/i3/config
+	ln -sfv ${BASEDIR}/i3/i3blocks.conf ~/.config/i3/i3blocks.conf
 
+	#bash
 
+	ln -sfv ${BASEDIR}/bash/bashrc ~/.bashrc
+	ln -sfv ${BASEDIR}/bash/bash_aliases ~/.bash_aliases
 
+	#X11
+
+	ln -svf ${BASEDIR}/X11/Xresources ~/.Xresources
+	xrdb -merge ~/.Xresources
+fi
